Smart_Inventory/backend/requirements.txt
flask
flask-cors
transformers
torch
Pillow

